Question 1
Answers I have worked on and answers from other that I have found online
for comparison.  They are each timed.  My first one is brute force and
takes by far the longest.  The online ones are pretty quick.  My second
one is the fastest at higher load by an order of magnitude but similar
to the online ones as the question was posed. The question asked for the
sum of multiples of 3 and 5 up to one thousand.  I made these up to on
million to see the performance difference.  Note that my second one can
do multiples for any two numbers by following a repeating pattern for
addition up to the lowest common multiple.