Fixed logic
Instead of starting from the lowest multiple of either number (the lower of the two numbers given) and going up to the multiple of the two numbers, I now start at the multiple of the two numbers and go to the next multiple. This may cut down on the number of iterations for lower numbers.

I'm also not sure if it works both ways, or if I actually fixed it. I remember thinking that the pattern was only between common multiples. Maybe I only got lucky with given parameters in the exercise. It seemed to happen last time when I labeled the lcm as just the product of the two numbers, which it happened to be but is the wrong method.

It has been a while since I did this and I was just noticing the LCM being the wrong calculation which made me go and refactor the whole thing, of course. I'm not even totally sure this works for everything, since I did not thoroughly test different combinations.